removed target tranformations, unified dataloaders, removed unused data augmentation functions
