+++ .gitignore
@@ -5,6 +5,7 @@ ASM_stereo_train_stage1/
 ASM_stereo_train_retrain1/
 KITTI_stage2/
 KITTI_train1k
+KITTI_train2k
 data/*
 temp/
 eigen/
