+++ misc/data_transforms.py
@@ -17,151 +17,11 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-import random, numbers
+import random
 import torch
 from torchvision import transforms
 from torchvision.transforms.functional import hflip
 import numpy as np
-from PIL import Image
-
-
-class Compose(object):
-    def __init__(self, co_transforms):
-        self.co_transforms = co_transforms
-
-    def __call__(self, input, target=None):
-        for t in self.co_transforms:
-            input, target = t(input, target)
-        return input, target
-
-
-class ArrayToTensor(object):
-    def __call__(self, array):
-        assert isinstance(array, np.ndarray)
-        if len(array.shape) == 3 and (array.shape[-1] == 3 or array.shape[-1] == 1):
-            array = np.moveaxis(array, -1, 0)
-        tensor = torch.from_numpy(array.copy())
-        return tensor.float()
-
-
-class RandomResizeCrop(object):
-    """Crops the given PIL.Image at a random location to have a region of
-    the given size. size can be a tuple (target_height, target_width)
-    or an integer, in which case the target will be of a square shape (size, size)
-    """
-
-    def __init__(self, size, down, up):
-        if isinstance(size, numbers.Number):
-            self.size = (int(size), int(size))
-        else:
-            self.size = size
-        self.s_factor = (down, up)
-
-    def __call__(self, inputs, targets=None):
-        h, w, _ = inputs[0].shape
-        th, tw = self.size
-
-        min_factor = max(
-            max((th + 1) / h, (tw + 1) / w), self.s_factor[0]
-        )  # plus one to ensure
-        max_factor = self.s_factor[1]
-        factor = np.random.uniform(low=min_factor, high=max_factor)
-
-        for i in range(len(inputs)):
-            input = Image.fromarray(inputs[i]).resize(
-                (int(w * factor), int(h * factor)), resample=Image.BICUBIC
-            )
-            inputs[i] = np.array(input)
-        if targets is not None:
-            for i in range(len(targets)):
-                target = Image.fromarray(targets[i]).resize(
-                    (int(w * factor), int(h * factor)), resample=Image.BICUBIC
-                )
-                targets[i] = np.array(target)
-
-        h, w, _ = inputs[0].shape
-        x1 = random.randint(0, w - tw)
-        y1 = random.randint(0, h - th)
-        for i in range(len(inputs)):
-            inputs[i] = inputs[i][y1 : y1 + th, x1 : x1 + tw]
-        if targets is not None:
-            for i in range(len(targets)):
-                targets[i] = targets[i][y1 : y1 + th, x1 : x1 + tw]
-        return inputs, targets
-
-
-class RandomHorizontalFlip(object):
-    """Randomly horizontally flips the given PIL.Image with a probability of 0.5
-    if doing this on disparity estimation you need both disparities left and right need stereo targets
-    """
-
-    def __init__(self):
-        return
-
-    def __call__(self, inputs, targets=None):
-        o_inputs = []
-
-        if random.random() < 0.5:
-            o_inputs.append(np.copy(np.fliplr(inputs[1])))
-            o_inputs.append(np.copy(np.fliplr(inputs[0])))
-            if targets is not None:
-                o_target = []
-                o_target.append(np.copy(np.fliplr(targets[1])))
-                o_target.append(np.copy(np.fliplr(targets[0])))
-                return o_inputs, o_target
-            else:
-                return o_inputs, targets
-        else:
-            return inputs, targets
-
-
-class RandomGamma(object):
-    def __init__(self, min=1, max=1):
-        self.min = min
-        self.max = max
-        self.A = 255
-
-    def __call__(self, inputs, targets=None):
-        if random.random() < 0.5:
-            factor = random.uniform(self.min, self.max)
-            for i in range(len(inputs)):
-                inputs[i] = self.A * ((inputs[i] / 255) ** factor)
-            return inputs, targets
-        else:
-            return inputs, targets
-
-
-class RandomBrightness(object):
-    def __init__(self, min=0, max=0):
-        self.min = min
-        self.max = max
-
-    def __call__(self, inputs, targets=None):
-        if random.random() < 0.5:
-            factor = random.uniform(self.min, self.max)
-            for i in range(len(inputs)):
-                inputs[i] = inputs[i] * factor
-                inputs[i][inputs[i] > 255] = 255
-            return inputs, targets
-        else:
-            return inputs, targets
-
-
-class RandomCBrightness(object):
-    def __init__(self, min=0, max=0):
-        self.min = min
-        self.max = max
-
-    def __call__(self, inputs, targets=None):
-        if random.random() < 0.5:
-            for i in range(len(inputs)):
-                for c in range(3):
-                    factor = random.uniform(self.min, self.max)
-                    inputs[i][:, :, c] = inputs[i][:, :, c] * factor
-                inputs[i][inputs[i] > 255] = 255
-            return inputs, targets
-        else:
-            return inputs, targets
 
 
 class NormalizeInverse(transforms.Normalize):
