+++ testing/test_k.py
@@ -42,17 +42,15 @@ from misc.postprocessing import ms_pp, local_normalization
 def main(args, device="cpu"):
     print("-------Testing on " + str(device) + "-------")
 
-    input_transform = transforms.Compose(
-        [
-            data_transforms.ArrayToTensor(),
-            transforms.Normalize(
-                mean=[0, 0, 0], std=[255, 255, 255]
-            ),  # (input - mean) / std
-            transforms.Normalize(mean=[0.411, 0.432, 0.45], std=[1, 1, 1]),
-        ]
+    # Set up data augmentations
+    input_transform = data_transforms.ApplyToMultiple(
+        transforms.Compose(
+            [
+                transforms.ToTensor(),
+                transforms.Normalize(mean=[0.411, 0.432, 0.45], std=[1, 1, 1]),
+            ]
+        )
     )
-
-    target_transform = transforms.Compose([data_transforms.ArrayToTensor()])
 
     # Torch Data Set List
     input_path = os.path.join(args.data_directory, args.dataset)
@@ -60,17 +58,14 @@ def main(args, device="cpu"):
         split=args.test_split,
         dataset=args.dataset,
         root=input_path,
-        disp=True,
         transform=input_transform,
-        target_transform=target_transform,
     )
 
     print("len(test_dataset)", len(test_dataset))
     # Torch Data Loader
-    args.batch_size = 1  # kitty mixes image sizes!
     val_loader = torch.utils.data.DataLoader(
         test_dataset,
-        batch_size=args.batch_size,
+        batch_size=1,
         num_workers=args.workers,
         pin_memory=False,
         shuffle=False,
